GP-1967: Fix creditor being reset when closing groups
When CiviSEPA closes transaction groups of type OOF or FRST, it
updates the mandate status to SENT or RCUR respectively. In
environments with multiple SEPA creditors, because the API does not
check whether it is creating a new mandate or updating an existing
one, this update causes the creditor to be set to the default
creditor.